Time: 0 ms (100.00%), Space: 6 MB (25.17%) - LeetHub

+++ 441-arranging-coins/441-arranging-coins.cpp
@@ -29,8 +29,8 @@ public:
         
         n%=INT_MAX;
        
-        long long int x1=(-1+sqrtl(1+(8*n)))/2;
-        long long int  x3=(-1-sqrtl(1+(8*n)))/2;
+        int x1=(-1+sqrt(1+(8*n)))/2;
+        int  x3=(-1-sqrt(1+(8*n)))/2;
         
         return x1>0?x1:x3;
